Revert "chore: Move function out of receivers"
This reverts commit b2bda96e89eab249ede1364ad0f12b0b9b186c63.
