+++ src/sentry/utils/db.py
@@ -1,28 +1,9 @@
-import logging
 from collections.abc import Sequence
 from contextlib import ExitStack
-from functools import wraps
 
 import sentry_sdk
 from django.db import DEFAULT_DB_ALIAS, connections, transaction
-from django.db.utils import OperationalError, ProgrammingError
 from sentry_sdk.integrations import Integration
-
-
-def handle_db_failure(func, using: str = "default", wrap_in_transaction: bool = True):
-    @wraps(func)
-    def wrapped(*args, **kwargs):
-        try:
-            if wrap_in_transaction:
-                with transaction.atomic(using):
-                    return func(*args, **kwargs)
-            else:
-                return func(*args, **kwargs)
-        except (ProgrammingError, OperationalError):
-            logging.exception("Failed processing signal %s", func.__name__)
-            return
-
-    return wrapped
 
 
 def atomic_transaction(
